+++ web/templates/version.html
@@ -1,2 +1,2 @@
 Kipa SVN {{"$Rev$"|cut:"$"}}
-{% comment %} 18:51:37 12/16/17 {% endcomment %}
+{% comment %} 17:11:13 12/17/17 {% endcomment %}
