+++ web/tupa/urls.py
@@ -44,7 +44,5 @@ urlpatterns = [
         url(r'^(?P<kisa_nimi>[^/]+)/tulosta/csv/sarja/(?P<sarja_id>\d+)/$', views.sarjanTuloksetCSV, name='sarjanTuloksetCSV'),
         url(r'^(?P<kisa_nimi>[^/]+)/tulosta/csv/$', views.tulosta, name='tulosta'),
         url(r'^(?P<kisa_nimi>[^/]+)/tulosta/piirit/$', views.piirit, name='piirit'),
-        url(r'^(?P<kisa_nimi>[^/]+)/kayttajat/$', views.kayttajat, name='kayttajat'),
-        url(r'^kayttaja/(?P<user_name>[^/]+)/$', views.kayttajan_tiedot, name='kayttaja'),
         ]
