+++ web/urls.py
@@ -12,6 +12,7 @@ urlpatterns = [
         url(r'^kipa/',  include('tupa.urls')),
         url(r'^admin/', admin.site.urls),
         url(r'^$', RedirectView.as_view(url='/kipa/', permanent=True)),
+        url(r'^kipa/', include('user_management.urls')),
 ]
 
 if settings.DEBUG :
