+++ web/user_management/apps.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.apps import AppConfig
+
+
+class UserManagementConfig(AppConfig):
+    name = 'user_management'
